Combine test scene to avoid huge copy paste
